+++ kroxylicious-filters/kroxylicious-encryption/src/main/java/io/kroxylicious/filter/encryption/Aad.java
@@ -7,15 +7,21 @@
 package io.kroxylicious.filter.encryption;
 
 /**
- * Enumerates the sets of metadata which can be used as additional authenticated date for AEAD ciphers.
+ * <p>Enumerates the sets of metadata which can be used as additional authenticated data (AAD) for AEAD ciphers.
  * Each element in this enumeration corresponds to a schema for the serialization of that set of metadata.
+ * </p>
+ *
+ * <p>The AAD available to be used depends on the {@link EncryptionVersion}:
+ * You cannot add a new cipher to an existing EncryptionVersion because doing
+ * so would expose old versions of the software to ciphers that they don't know
+ * about meaning they won't be able to supprt the decryption guarantee.</p>
  */
 public enum Aad {
     /**
      * No AAD
      */
     NONE((byte) 0, EncryptionVersion.V1);
-    // TODO we need to constraint which Aad options are allowed in which EncrytionVersions
+
     // /**
     // * AAD consisting of the batch metadata plus the records position within the batch.
     // */
