+++ kroxylicious-filters/kroxylicious-encryption/src/main/java/io/kroxylicious/filter/encryption/EncryptionConfigurationException.java
@@ -6,6 +6,10 @@
 
 package io.kroxylicious.filter.encryption;
 
+/**
+ * Represents a misconfiguration of encryption.
+ * Exceptions of this type mean the proxy admin has messed up.
+ */
 public class EncryptionConfigurationException extends EncryptionException {
 
     public EncryptionConfigurationException(String message) {
