+++ kroxylicious-filters/kroxylicious-encryption/src/main/java/io/kroxylicious/filter/encryption/inband/ExhaustedDekException.java
@@ -8,6 +8,11 @@ package io.kroxylicious.filter.encryption.inband;
 
 import io.kroxylicious.filter.encryption.EncryptionException;
 
+/**
+ * Exception indicating that a DEK has been used for
+ * encryption too many times.
+ * In theory this should never be thrown in practice.
+ */
 public class ExhaustedDekException extends EncryptionException {
     public ExhaustedDekException(String msg) {
         super(msg);
